Modularize IMAP block by extracting common utilities and simplifying trigger logic
