+++ packages/blocks/imap/src/lib/common/imap-auth.ts
@@ -1,0 +1,48 @@
+import { BlockAuth, Property } from '@openops/blocks-framework';
+import { buildClient } from './build-client';
+
+export const imapAuth = BlockAuth.CustomAuth({
+  authProviderKey: 'imap',
+  authProviderDisplayName: 'IMAP',
+  authProviderLogoUrl: 'https://static.openops.com/blocks/imap.png',
+  required: true,
+  props: {
+    host: Property.ShortText({
+      displayName: 'Host',
+      required: true,
+    }),
+    username: Property.ShortText({
+      displayName: 'Username',
+      required: true,
+    }),
+    password: Property.SecretText({
+      displayName: 'Password',
+      required: true,
+    }),
+    port: Property.Number({
+      displayName: 'Port',
+      required: true,
+      defaultValue: 143,
+    }),
+    tls: Property.Checkbox({
+      displayName: 'Use TLS',
+      defaultValue: false,
+      required: true,
+    }),
+  },
+  validate: async ({ auth }) => {
+    const imapClient = buildClient(auth);
+    try {
+      await imapClient.connect();
+      await imapClient.noop();
+      return { valid: true };
+    } catch (e) {
+      return {
+        valid: false,
+        error: JSON.stringify(e),
+      };
+    } finally {
+      await imapClient.logout();
+    }
+  },
+});
